feat(ts-client): deprecate previous client
still exporting it for next.js example
