+++ packages/ts-client/_src/classes/index.ts
@@ -1,0 +1,3 @@
+export * from './Client.class'
+export * from './Entity.class'
+export * from './Operation.class'
