+++ packages/ts-client/_src/index.ts
@@ -1,0 +1,2 @@
+export * from './classes/Client.class'
+export * from './types'
