+++ packages/ts-client/_src/types/Beacon.types.ts
@@ -1,0 +1,4 @@
+export interface Beacon {
+  beacon: string
+  __typename: 'Beacon'
+}
