+++ packages/ts-client/_src/types/Fragments.types.ts
@@ -1,0 +1,20 @@
+export interface TocItem {
+  tag: string
+  blockIndex: number
+  href: string
+  title: string
+  level: number
+}
+
+export interface MentionItem {
+  name: string
+  emailAddress: string
+}
+
+export interface FootnoteItem {
+  id: string
+  index: number
+  refId: string
+  refValue: string
+  valueHTML: string
+}
