+++ packages/ts-client/_src/types/UnbodyDocumentTypeNames.types.ts
@@ -1,0 +1,7 @@
+export enum UnbodyDocumentTypeNames {
+  GoogleDoc = 'GoogleDoc',
+  GoogleCalendarEvent = 'GoogleCalendarEvent',
+  TextBlock = 'TextBlock',
+  ImageBlock = 'ImageBlock',
+  AudioFile = 'AudioFile',
+}
