+++ packages/ts-client/_src/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './query.utils'
+export * from './WhereOperands.class'
