+++ packages/ts-client/index.ts
@@ -1,3 +1,3 @@
-export * from './src/classes/Client'
-export * from './src/types'
-export * from './src/utils'
+export * from './_src/classes'
+export * from './_src/types'
+export * from './_src/utils'
