--- packages/ts-client/src/classes/index.ts
@@ -1,4 +1,0 @@
-export * from './Client.class';
-export * from './Entity.class';
-export * from './Operation.class';
-
