--- packages/ts-client/src/types/Beacon.types.ts
@@ -1,4 +1,0 @@
-export interface Beacon {
-    beacon: string
-    __typename: 'Beacon'
-}
