--- packages/ts-client/src/types/UnbodyDocumentTypeNames.types.ts
@@ -1,7 +1,0 @@
-export enum UnbodyDocumentTypeNames {
-    GoogleDoc = 'GoogleDoc',
-    GoogleCalendarEvent = 'GoogleCalendarEvent',
-    TextBlock = 'TextBlock',
-    ImageBlock = 'ImageBlock',
-    AudioFile = 'AudioFile',
-}
