--- packages/ts-client/src/types/index.ts
@@ -1,12 +1,0 @@
-export * from './UnbodyDocumentTypeNames.types';
-export * from './AdditionalProps.types';
-export * from './BaseObject.types';
-export * from './Beacon.types';
-export * from './Fragments.types';
-export * from './ImageBlock.types';
-export * from './TextBlock.types';
-export * from './AudioFile.types';
-export * from './GoogleDoc.types';
-export * from './GoogleCalendarEvent.types';
-export * from './Filters.types';
-export * from './api.types';
